+++ src/domUpdates.js
@@ -1,3 +1,5 @@
+import Recipe from './recipe';
+
 let favButton = document.querySelector('.view-favorites');
 let homeButton = document.querySelector('.home')
 let cardArea = document.querySelector('.all-cards');
@@ -19,21 +21,21 @@ let domUpdates = {
     }
     recipes.forEach(recipe => {
       cardArea.insertAdjacentHTML('afterbegin',
-        `<div id='${recipe.id}' class='card'>
-            <header id='${recipe.id}' class='card-header'>
+        `<div data-id='${recipe.id}' class='card'>
+            <header data-id='${recipe.id}' class='card-header'>
                 <label for='add-button' class='hidden'>Click to add recipe</label>
-                <button id='${recipe.id}' aria-label='add-button' class='add-button card-button'>
-                <img id='${recipe.id} favorite' class='add'
+                <button data-id='${recipe.id}' aria-label='add-button' class='add-button card-button'>
+                <img data-id='${recipe.id}' favorite' class='add'
                 src='https://image.flaticon.com/icons/svg/32/32339.svg' alt='Add to
                 recipes to cook'>
                 </button>
                 <label for='favorite-button' class='hidden'>Click to favorite recipe
                 </label>
-                <button id='${recipe.id}' aria-label='favorite-button' class='favorite favorite${recipe.id} card-button'></button>
+                <button data-id='${recipe.id}' aria-label='favorite-button' class='favorite favorite${recipe.id} card-button'></button>
             </header>
-                <img id='${recipe.id}' tabindex='0' class='card-picture'
+                <img data-id='${recipe.id}' tabindex='0' class='card-picture'
                 src='${recipe.image}' alt='click to view recipe for ${recipe.name}'>
-                <span id='${recipe.id}' class='recipe-name'>${recipe.name}</span>
+                <span data-id='${recipe.id}' class='recipe-name'>${recipe.name}</span>
         </div>`)
     })
     this.getFavorites(user);
@@ -57,21 +59,21 @@ let domUpdates = {
       favButton.innerHTML = 'Refresh Favorites'
       cardArea.innerHTML = '';
       user.favoriteRecipes.forEach(recipe => {
-        cardArea.insertAdjacentHTML('afterbegin', `<div id='${recipe.id}'
+        cardArea.insertAdjacentHTML('afterbegin', `<div data-id='${recipe.id}'
             class='card'>
-            <header id='${recipe.id}' class='card-header'>
+            <header data-id='${recipe.id}' class='card-header'>
             <label for='add-button' class='hidden'>Click to add recipe</label>
-            <button id='${recipe.id}' aria-label='add-button' class='add-button card-button'>
-            <img id='${recipe.id}' class='add'
+            <button data-id='${recipe.id}' aria-label='add-button' class='add-button card-button'>
+            <img data-id='${recipe.id}' class='add'
             src='https://image.flaticon.com/icons/svg/32/32339.svg' alt='Add to
             recipes to cook'></button>
             <label for='favorite-button' class='hidden'>Click to favorite recipe
             </label>
-            <button id='${recipe.id}' aria-label='favorite-button' class='favorite favorite-active card-button'>
+            <button data-id='${recipe.id}' aria-label='favorite-button' class='favorite favorite-active card-button'>
             </button></header>
-            <img id='${recipe.id}' tabindex='0' class='card-picture'
+            <img data-id='${recipe.id}' tabindex='0' class='card-picture'
             src='${recipe.image}' alt='Food from recipe'>
-            <span id='${recipe.id}' class='recipe-name'>${recipe.name}</span>
+            <span data-id='${recipe.id}' class='recipe-name'>${recipe.name}</span>
 
             </div>`)
       })
@@ -79,7 +81,7 @@ let domUpdates = {
   },
   favoriteCard(event, cookbook, user) {
     let specificRecipe = cookbook.recipes.find(recipe => {
-      if (recipe.id === Number(event.target.id)) {
+      if (recipe.id === Number(event.target.dataset.id)) {
         return recipe;
       }
     })
@@ -113,23 +115,23 @@ let domUpdates = {
       recipiesSpan.insertAdjacentHTML('afterbegin', `<p>No recipies to Cook!</p>`)
     } else {
       user.recipesToCook.forEach(recipe => {
-        recipiesSpan.insertAdjacentHTML('afterbegin', `<div id='${recipe.id}'
+        recipiesSpan.insertAdjacentHTML('afterbegin', `<div data-id='${recipe.id}'
                 class='card'>
-                    <header id='${recipe.id}' class='card-header'>
+                    <header data-id='${recipe.id}' class='card-header'>
                         <label for='add-button' class='hidden'>Click to add recipe</label>
-                        <button id='${recipe.id}' aria-label='add-button' class='add-button card-button'>
-                        <img id='${recipe.id} favorite' class='add'
+                        <button data-id='${recipe.id}' aria-label='add-button' class='add-button card-button'>
+                        <img data-id='${recipe.id}' id='${recipe.id} favorite' class='add'
                         src='https://image.flaticon.com/icons/svg/32/32339.svg' alt='Add to
                         recipes to cook'>
                         </button>
                         <label for='favorite-button' class='hidden'>Click to favorite recipe
                         </label>
-                        <button id='${recipe.id}' aria-label='favorite-button' class='favorite favorite${recipe.id} card-button'></button>
+                        <button data-id='${recipe.id}' aria-label='favorite-button' class='favorite favorite${recipe.id} card-button'></button>
                     </header>
-                        <span id='${recipe.id}' class='recipe-name'>${recipe.name}</span>
-                        <img id='${recipe.id}' tabindex='0' class='card-picture'
+                        <span data-id='${recipe.id}' class='recipe-name'>${recipe.name}</span>
+                        <img data-id='${recipe.id}' tabindex='0' class='card-picture'
                         src='${recipe.image}' alt='click to view recipe for ${recipe.name}'>
-                        <button id='${recipe.id}' class='cook-meal navButton'>Cook this Meal</button>
+                        <button data-id='${recipe.id}' class='cook-meal navButton'>Cook this Meal</button>
                 </div>`);
       });
       this.getFavorites(user);
@@ -151,35 +153,66 @@ let domUpdates = {
     cantCookSpan.innerHTML = `Cooked ${recipe.name}!`;
   },
   searchBarSearch(recipeRepository, ingredientsData) {
-    console.log("I'm here");
     let userRequestSearch = searchInput.value;
     if (cardArea.classList.contains('all')) {
       cardArea.classList.remove('all')
     } else {
       cardArea.innerHTML = '';
       const searchResults = recipeRepository.retrieveListByNameOrIngredients(userRequestSearch, ingredientsData)
-      //console.log(searchResults);
       searchResults.forEach(recipe => {
 
-        cardArea.insertAdjacentHTML('afterbegin', `<div id='${recipe.id}'
+        cardArea.insertAdjacentHTML('afterbegin', `<div data-id='${recipe.id}' 
             class='card'>
-            <header id='${recipe.id}' class='card-header'>
+            <header data-id='${recipe.id}' class='card-header'>
             <label for='add-button' class='hidden'>Click to add recipe</label>
-            <button id='${recipe.id}' aria-label='add-to-pantry-button' class='add-button card-button'>
-            <img id='${recipe.id}' class='add'
+            <button data-id='${recipe.id}' aria-label='add-to-pantry-button' class='add-button card-button'>
+            <img data-id='${recipe.id}' class='add'
             src='https://image.flaticon.com/icons/svg/32/32339.svg' alt='Add to
             recipes to cook'></button>
             <label for='favorite-button' class='hidden'>Click to favorite recipe
             </label>
-            <button id='${recipe.id}' aria-label='favorite-button' class='favorite card-button'>
+            <button data-id='${recipe.id}' aria-label='favorite-button' class='favorite card-button'>
             </button></header>
-            <img id='${recipe.id}' tabindex='0' class='card-picture'
+            <img data-id='${recipe.id}' tabindex='0' class='card-picture'
             src='${recipe.image}' alt='Food from recipe'>
-            <span id='${recipe.id}' class='recipe-name'>${recipe.name}</span>
+            <span data-id='${recipe.id}' class='recipe-name'>${recipe.name}</span>
             </div>`)
       })
     }
-
+  },
+  displayDirections(event, cookbook, globalIngredientsData, convertToName) {
+    let newRecipeInfo = cookbook.recipes.find(recipe => {
+      if (recipe.id === Number(event.target.dataset.id)) {
+        return recipe;
+      }
+    })
+    let recipeObject = new Recipe(newRecipeInfo, globalIngredientsData);
+    const ingrdientWithName = convertToName(recipeObject);
+    recipeObject.ingredients = ingrdientWithName;
+    let cost = recipeObject.calculateCost()
+    let costInDollars = (cost / 100).toFixed(2)
+    cardArea.classList.add('all');
+    cardArea.innerHTML = `<h3>${recipeObject.name}</h3>
+    <p class='all-recipe-info'>
+    <strong>It will cost: </strong><span class='cost recipe-info'>
+    $${costInDollars}</span><br><br>
+    <strong>You will need: </strong><span class='ingredients recipe-info'></span>
+    <strong>Instructions: </strong><ol><span class='instructions recipe-info'>
+    </span></ol>
+    </p>`;
+    let ingredientsSpan = document.querySelector('.ingredients');
+    let instructionsSpan = document.querySelector('.instructions');
+    recipeObject.ingredients.forEach(ingredient => {
+      ingredientsSpan.insertAdjacentHTML('afterbegin', `<ul><li>
+      ${ingredient.quantity.amount.toFixed(2)} ${ingredient.quantity.unit}
+      ${ingredient.name}</li></ul>
+      `)
+    })
+    recipeObject.instructions.forEach(instruction => {
+      instructionsSpan.insertAdjacentHTML('beforebegin', `<li>
+      ${instruction.instruction}</li>
+      `)
+    })
   }
 };
